requirements.txt
pandas==2.1.4
plotly==5.18.0
pytest==7.4.4
requests==2.31.0
ipykernel==6.28.0
nbformat==5.9.2
matplotlib==3.8.2
python-dotenv==1.0.0